moved introspection functions to Auth folder
