+++ src/index.js
@@ -5,11 +5,11 @@ const Query = require("./resolvers/Query");
 const Mutation = require("./resolvers/Mutations");
 const {PhoneNumber} = require("./resolvers/Scalars");
 const config = require("./config");
-const AuthDirectives = require('./Auth/Directives');
+const AuthDirectives = require("./Auth/Directives");
 const fs = require("fs");
 const { connectMessageQueueListener } = require("./Service_Mesh/listener_connector");
 const { connectMessageQueuePublisher } = require("./Service_Mesh/publisher_connector");
-const introspect = require("./introspection");
+const introspect = require("./Auth/introspection");
 
 const resolvers = {
   Query,
